requirements.txt
langchain
langchain_community
langchain_openai
python-dotenv
chromadb
langchain_chroma